+++ TEditXna/Terraria/World.Settings.cs
@@ -186,7 +186,7 @@ namespace TEditXNA.Terraria
                                         Name = curFrame.Name + ", " + curFrame.Variety,
                                         Origin = curFrame.UV,
                                         Size = curTile.FrameSize,
-                                        Tile = (byte)curTile.Id,
+                                        Tile = (ushort)curTile.Id,
                                         TileName = curTile.Name
                                     });
                 }
